// src/utils/types.ts
export interface Particle {
  id: number;
  position: Vec3;
  offset: Vec3;
  count: number;
  movementSpeed: number;
  longDistanceRender: boolean;
}